Lots of changes, particularly to modifications
Fixed a bug with modifying the orientation of existing elements
There was a bug with sending large amounts of data from GH to Revit,
limiting the number of created elements to a few hundred at best.  I
believe this is resolved and I have successfully sent thousands of
elements.
The command is now a split button to add additional commands
The toggle is the primary/default button
A button to access settings were added.  It's now possible to disable it
in one session and enable it in another session if necessary.
Creating elements now stores both the GH component instance GUID and a
run id (sequential integer).  You can use this to create or modify
different sets of data with the same component.
With this new run id, the modification pop up has changed to a new form.
It lets you select a matching run from the combobox and then choose to
modify, create a new run, or cancel.