Changes so that the GH component nickname gets sent to Revit
To make it easier to potentially differentiate between what component
sent what data, I've added the nickname from GH to be included with the
schema created to store the GH information.  So now instead of just the
GH InstanceGUID you also get the nickname.  Particularly useful when
using the Select Run command in Revit.

NOTE: You can't modify a schema or the fields once it's created so for
now it's not overwriting if you change the nickname.  Only the initial
nickname is held.