ElementId Updates
Adding updates that will extract elements that can be used with
ElementId parameters (materials, levels, phases, images) and a starting
point for being able to create levels and grids with Lyrebird.

Fixed some issues with the Revit 2015 buttons (settings, select) that
were not working because of an issue with finding an image.